# Distributed.Services/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Jobag",
    "version": "v1"
  },
  "paths": {
    "/api/Company/Register": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Permite registrar una compañía",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterCompanyDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CompanyDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
    "/api/Company/Login": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Permite a una compañía ingresar al sistema",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginCompanyDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/LoginCompanyResponseDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
    "/api/Company/Get": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Permite obtener una compañía por id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "companyId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CompanyDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
...
    "/api/JobOffer/Get": {
      "get": {
        "tags": [
          "JobOffer"
        ],
        "summary": "Permite obtener una oferta de trabajo por id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "jobOfferId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobOfferDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
    "/api/JobOffer/GetAll": {
      "get": {
        "tags": [
          "JobOffer"
        ],
        "summary": "Permite obtener todas las ofertas de trabajo",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobOfferDtoListJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
    "/api/JobOffer/GetByCompany": {
      "get": {
        "tags": [
          "JobOffer"
        ],
        "summary": "Permite obtener todas las ofertas de trabajo de una compañía",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobOfferDtoListJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
...
    "/api/Postulant/Register": {
      "post": {
        "tags": [
          "Postulant"
        ],
        "summary": "Permite registrar un postulante",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterPostulantDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PostulantDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
    "/api/Postulant/Login": {
      "post": {
        "tags": [
          "Postulant"
        ],
        "summary": "Permite a un postulante ingresar al sistema",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginPostulantDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/LoginPostulantResponseDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
    "/api/Postulant/Get": {
      "get": {
        "tags": [
          "Postulant"
        ],
        "summary": "Permite obtener un postulante por id",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "postulantId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PostulantDtoJsonResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/StringJsonResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CompanyDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "companyName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "websiteUrl": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "bannerPicture": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "passwordHash": {
          "format": "byte",
          "type": "string"
        },
        "passwordSalt": {
          "format": "byte",
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CompanyDtoJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/CompanyDto"
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
...
    "JobOfferDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "wage": {
          "format": "double",
          "type": "number"
        },
        "snippet": {
          "type": "string"
        },
        "bannerPicture": {
          "type": "string"
        },
        "hasHired": {
          "type": "boolean"
        },
        "companyId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "JobOfferDtoJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/JobOfferDto"
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
    "JobOfferDtoListJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JobOfferDto"
          }
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
    "LoginCompanyDto": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "LoginCompanyResponseDto": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "company": {
          "$ref": "#/definitions/CompanyDto"
        }
      }
    },
    "LoginCompanyResponseDtoJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/LoginCompanyResponseDto"
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
    "LoginPostulantDto": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "LoginPostulantResponseDto": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "postulant": {
          "$ref": "#/definitions/PostulantDto"
        }
      }
    },
    "LoginPostulantResponseDtoJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/LoginPostulantResponseDto"
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
    "PostulantDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "salary": {
          "format": "double",
          "type": "number"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "format": "int32",
          "type": "integer"
        },
        "workingPlaces": {
          "type": "string"
        },
        "languages": {
          "type": "string"
        },
        "skills": {
          "type": "string"
        },
        "studyCenter": {
          "type": "string"
        },
        "about": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "bannerPicture": {
          "type": "string"
        },
        "passwordHash": {
          "format": "byte",
          "type": "string"
        },
        "passwordSalt": {
          "format": "byte",
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "PostulantDtoJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/PostulantDto"
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
...
    "RegisterCompanyDto": {
      "required": [
        "confirmPassword",
        "email",
        "password",
        "roleId"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 4,
          "type": "string"
        },
        "confirmPassword": {
          "minLength": 1,
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "RegisterPostulantDto": {
      "required": [
        "confirmPassword",
        "email",
        "password",
        "roleId"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 4,
          "type": "string"
        },
        "confirmPassword": {
          "minLength": 1,
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "StringJsonResult": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "string"
        },
        "warning": {
          "type": "boolean"
        }
      }
    },
...
  }
}